Abstractify test scene
I think in this case it's genuinely reasonable to use abstracts to
reduce boilerplate.
